refactor token functions into token.py
